+++ service/src/java/org/apache/hive/service/cli/thrift/ThriftCLIService.java
@@ -73,7 +73,7 @@ public abstract class ThriftCLIService extends AbstractService implements TCLISe
 
   protected int minWorkerThreads;
   protected int maxWorkerThreads;
-  protected int workerKeepAliveTime;
+  protected long workerKeepAliveTime;
 
   protected static HiveAuthFactory hiveAuthFactory;
